working with journal
This version is tested to work with journal. The decision to run
kas_alias for modules in a single shot stems in the observation that
kas_alias has a non-negligible amount of work to do at start, and
that, in case of a run for each module, this work is replicated for
each module it processes. Despite if seeing this from a single thread
application point of view, the consideration this approach makes the
performances better stands, the consideration that Makefile can
instantiate a process for each processor obsoletes this aspect. The
current PoC is designed to work as single thread one shot, might be
worth considering, add a feature to have a version that operates on a
single module rather than all at once. This version of journaling
system is there to allow the continuation of interrupted builds. It
allows application to discard already processed modules to be
corrupted by an additional application of aliases. The journaling has
been chosen to avoid to rerun nm on modules before Processing them as
it would be very expansive job. This commit is the starting point for
a new kas_alias action to allow a single module processing. New
version should have:
* new "single_module" action to process just one module.
* journaling system made optional and moved on action specific flags
* pre processed nm data from core image, in single_module should be made
optional, if used can be used as cache to speedup things, but they also
intro duce a light risk of corruption (double aliasese).
for this reason it is worth considering an optional usage of the journal

Signed-off-by: Alessandro Carminati (Red Hat) <alessandro.carminati@gmail.com>